Renamed files for graph visualization interface
