+++ src/chart/graph/GraphChartVisualization2D.tsx
@@ -2,7 +2,7 @@ import React, { useRef } from 'react';
 import ForceGraph2D from 'react-force-graph-2d';
 import { executeActionRule, getRuleWithFieldPropertyName } from '../../extensions/advancedcharts/Utils';
 import { getTooltip } from './component/GraphChartTooltip';
-import { GraphChartVisualizationProps } from './GraphChartVisualizationProps';
+import { GraphChartVisualizationProps } from './GraphChartVisualization';
 import { generateNodeCanvasObject } from './util/NodeUtils';
 import { generateRelCanvasObject } from './util/RelUtils';
