+++ src/chart/graph/component/button/GraphChartFitViewButton.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Tooltip } from '@mui/material';
-import { GraphChartVisualizationProps } from '../../GraphChartVisualizationProps';
+import { GraphChartVisualizationProps } from '../../GraphChartVisualization';
 import { IconButton } from '@neo4j-ndl/react';
 import { FitToScreenIcon } from '@neo4j-ndl/react/icons';
